Fix occasional hang at end of NIS server stream.
If the loop is reading fast enough, the buffer will always be near empty
on each iteration. This can cause the EOF sequence to be split across
multiple iterations of the loop, causing it to go undetected. To fix
this we instead check the full output, instead of just the data that was
read on the current loop iteration.